Add files via upload
These are the files that run the app, however, there is also Flask and Ollama running in the background to facilitate the running.  The architecture is as follows, I have a Flask server for the webapp, Ollama to orchestrate the LLM and Flask, finally I am using Llama 3.2 as the LLM model to discuss with the user.